+++ 17_PhuongDong_12_HienDuy/frmKhachHang.cs
@@ -142,27 +142,39 @@ namespace _17_PhuongDong_12_HienDuy
             Xuly_Textbox(true);
             Xuly_Chucnang(true);
             string sql = "";
-            if(txtDiaChi.Text=="" || txtHoTen.Text=="" || txtMaKH.Text=="" || txtSoCMND.Text=="" || txtSoDienThoai.Text=="" || cbmGioiTinh.SelectedIndex==-1  )
+            if (txtDiaChi.Text == "" || txtHoTen.Text == "" || txtMaKH.Text == "" || txtSoCMND.Text == "" || txtSoDienThoai.Text == "" || cbmGioiTinh.SelectedIndex == -1)
             {
                 MessageBox.Show("Vui lòng điền đầy đủ thông tin!", "Thông báo");
             }
-            else 
-            { 
-            if (flag == 1)
-            {
-                sql = "insert into KhachHang values ('" + txtMaKH.Text +"',N'"+ txtHoTen.Text + "','"+ dtpNgaySinh.Text +"','" + txtSoDienThoai.Text +"', '" + txtSoCMND.Text+"',N'" +txtDiaChi.Text+"',N'"+ cbmGioiTinh.Items[cbmGioiTinh.SelectedIndex] +"','" + dtpNgayDen.Text+"')";
-                MessageBox.Show("Bạn có muốn thêm?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            }
-            else if (flag == 2)
-            {
-                sql = "update KhachHang set MaKH = '" + txtMaKH.Text + "', HoTen = N'" + txtHoTen.Text + "', NgaySinh = '"+dtpNgaySinh.Text+"', SDT = '" + txtSoDienThoai.Text + "', CMND = '" + txtSoCMND.Text + "',DiaChi= N'" + txtDiaChi.Text + "', GioiTinh = N'" + cbmGioiTinh.Items[cbmGioiTinh.SelectedIndex] + "', NgayDen = '" + dtpNgayDen.Text + "' where MaKH = '" + txtMaKH.Text + "'";
-                MessageBox.Show("Bạn có muốn sửa thông tin khách hàng?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            }
             else
             {
-                sql = "delete from KhachHang where MaKH = '" + txtMaKH.Text + "'";
-                MessageBox.Show("Bạn có muốn xóa thông tin khách hàng?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            }
+                if (flag == 1)
+                {
+                    DialogResult kq = MessageBox.Show("Bạn có muốn thêm?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    if (kq == DialogResult.Yes)
+                    {
+                        sql = "insert into KhachHang values ('" + txtMaKH.Text + "',N'" + txtHoTen.Text + "','" + dtpNgaySinh.Text + "','" + txtSoDienThoai.Text + "', '" + txtSoCMND.Text + "',N'" + txtDiaChi.Text + "',N'" + cbmGioiTinh.Items[cbmGioiTinh.SelectedIndex] + "','" + dtpNgayDen.Text + "')";
+                    }
+                }
+                else if (flag == 2)
+                {
+                    DialogResult kq = MessageBox.Show("Bạn có muốn sửa thông tin khách hàng?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    if (kq == DialogResult.Yes)
+                    {
+                        sql = "update KhachHang set MaKH = '" + txtMaKH.Text + "', HoTen = N'" + txtHoTen.Text + "', NgaySinh = '" + dtpNgaySinh.Text + "', SDT = '" + txtSoDienThoai.Text + "', CMND = '" + txtSoCMND.Text + "',DiaChi= N'" + txtDiaChi.Text + "', GioiTinh = N'" + cbmGioiTinh.Items[cbmGioiTinh.SelectedIndex] + "', NgayDen = '" + dtpNgayDen.Text + "' where MaKH = '" + txtMaKH.Text + "'";
+                    }
+                }
+                else
+                {
+                    DialogResult kq = MessageBox.Show("Bạn có muốn xóa thông tin khách hàng?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    if (kq == DialogResult.Yes)
+                    {
+                        sql = "delete from KhachHang where MaKH = '" + txtMaKH.Text + "'";
+                    }
+
+
+                }
+                if (sql != "")
             if (c.CapNhatDuLieu(sql) != 0)
             {
                 KH = txtMaKH.Text;
